Updated QLearning
It works now, but slowdown is really poor: possibly due to iteration, qtable findingStates? Need to also fix resets when screen transitions, as it doesn't reset when screen transitions. Also need to check save/loads, and also check to see if there is a way of simplifying actions. Don't forget to eventually add iterative learning too :)